Set BUILD_USERNAME / BUILD_HOSTNAME to default values
 BUILD_USERNAME=android
 BUILD_HOSTNAME=localhost

This information does not really matter, in the contrary, it makes the
build simply more unique in regards to tracking. This is none of anybody's
business!
Android Q seems to have acknowledged this, as it has been made very easy to
define these values and the documentation talks a similar language.